+++ lisp/eab-desktop.el
@@ -1,6 +1,6 @@
 ;;; eab-desktop.el ---  eab desktop extension
 
-;; Copyright (C) 2010-2023 Evgeny Boykov
+;; Copyright (C) 2010-2024 Evgeny Boykov
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: