+++ lisp/eab-ido.el
@@ -1,6 +1,6 @@
 ;;; eab-ido.el --- eab ido extension
 
-;; Copyright (C) 2010-2023 Evgeny Boykov
+;; Copyright (C) 2010-2024 Evgeny Boykov
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: