+++ lisp/eab-kbd-ext.el
@@ -1,6 +1,6 @@
 ;;; eab-kbd-ext.el --- 
 
-;; Copyright (C) 2010-2023 Evgeny Boykov
+;; Copyright (C) 2010-2024 Evgeny Boykov
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: